refactor: clean up definitions in `lib.rs`, split out `SessionBuilder`
