+++ src/environment.rs
@@ -1,21 +1,22 @@
 use std::{
 	cell::UnsafeCell,
-	ffi::CString,
+	ffi::{self, CStr, CString},
+	ptr,
 	sync::{
 		atomic::{AtomicPtr, Ordering},
 		Arc
 	}
 };
 
-use tracing::debug;
+use ort_sys::c_char;
+use tracing::{debug, Level};
 
-use super::{
-	custom_logger,
-	error::{Error, Result},
-	ortsys, ExecutionProviderDispatch
-};
 #[cfg(feature = "load-dynamic")]
 use crate::G_ORT_DYLIB_PATH;
+use crate::{
+	error::{Error, Result},
+	extern_system_fn, ortsys, ExecutionProviderDispatch
+};
 
 struct EnvironmentSingleton {
 	cell: UnsafeCell<Option<Arc<Environment>>>
@@ -215,6 +216,57 @@ pub fn init() -> EnvironmentBuilder {
 pub fn init_from(path: impl ToString) -> EnvironmentBuilder {
 	let _ = G_ORT_DYLIB_PATH.set(Arc::new(path.to_string()));
 	EnvironmentBuilder::default()
+}
+
+/// ONNX's logger sends the code location where the log occurred, which will be parsed into this struct.
+#[derive(Debug)]
+struct CodeLocation<'a> {
+	file: &'a str,
+	line: &'a str,
+	function: &'a str
+}
+
+impl<'a> From<&'a str> for CodeLocation<'a> {
+	fn from(code_location: &'a str) -> Self {
+		let mut splitter = code_location.split(' ');
+		let file_and_line = splitter.next().unwrap_or("<unknown file>:<unknown line>");
+		let function = splitter.next().unwrap_or("<unknown function>");
+		let mut file_and_line_splitter = file_and_line.split(':');
+		let file = file_and_line_splitter.next().unwrap_or("<unknown file>");
+		let line = file_and_line_splitter.next().unwrap_or("<unknown line>");
+
+		CodeLocation { file, line, function }
+	}
+}
+
+extern_system_fn! {
+	/// Callback from C that will handle ONNX logging, forwarding ONNX's logs to the `tracing` crate.
+	pub(crate) fn custom_logger(_params: *mut ffi::c_void, severity: ort_sys::OrtLoggingLevel, category: *const c_char, _: *const c_char, code_location: *const c_char, message: *const c_char) {
+		assert_ne!(category, ptr::null());
+		let category = unsafe { CStr::from_ptr(category) };
+		assert_ne!(code_location, ptr::null());
+		let code_location_str = unsafe { CStr::from_ptr(code_location) }.to_str().unwrap();
+		assert_ne!(message, ptr::null());
+		let message = unsafe { CStr::from_ptr(message) }.to_str().unwrap();
+
+		let code_location = CodeLocation::from(code_location_str);
+		let span = tracing::span!(
+			Level::TRACE,
+			"ort",
+			category = category.to_str().unwrap_or("<unknown>"),
+			file = code_location.file,
+			line = code_location.line,
+			function = code_location.function
+		);
+
+		match severity {
+			ort_sys::OrtLoggingLevel::ORT_LOGGING_LEVEL_VERBOSE => tracing::event!(parent: &span, Level::TRACE, "{message}"),
+			ort_sys::OrtLoggingLevel::ORT_LOGGING_LEVEL_INFO => tracing::event!(parent: &span, Level::DEBUG, "{message}"),
+			ort_sys::OrtLoggingLevel::ORT_LOGGING_LEVEL_WARNING => tracing::event!(parent: &span, Level::INFO, "{message}"),
+			ort_sys::OrtLoggingLevel::ORT_LOGGING_LEVEL_ERROR => tracing::event!(parent: &span, Level::WARN, "{message}"),
+			ort_sys::OrtLoggingLevel::ORT_LOGGING_LEVEL_FATAL=> tracing::event!(parent: &span, Level::ERROR, "{message}")
+		}
+	}
 }
 
 #[cfg(test)]
