+++ src/session/mod.rs
@@ -1,463 +1,27 @@
 //! Contains the [`Session`] and [`SessionBuilder`] types for managing ONNX Runtime sessions and performing inference.
 
-#[cfg(not(target_family = "windows"))]
-use std::os::unix::ffi::OsStrExt;
-#[cfg(target_family = "windows")]
-use std::os::windows::ffi::OsStrExt;
-#[cfg(feature = "fetch-models")]
-use std::path::PathBuf;
-use std::{
-	ffi::CString,
-	marker::PhantomData,
-	ops::Deref,
-	os::raw::c_char,
-	path::Path,
-	ptr::{self, NonNull},
-	rc::Rc,
-	sync::{atomic::Ordering, Arc}
-};
+use std::{ffi::CString, marker::PhantomData, ops::Deref, os::raw::c_char, ptr::NonNull, sync::Arc};
 
-#[cfg(feature = "fetch-models")]
-use super::error::FetchModelError;
 use super::{
 	char_p_to_string,
-	environment::get_environment,
+	environment::Environment,
 	error::{assert_non_null_pointer, assert_null_pointer, status_to_result, Error, ErrorInternal, Result},
-	execution_providers::{apply_execution_providers, ExecutionProviderDispatch},
 	extern_system_fn,
 	io_binding::IoBinding,
 	memory::Allocator,
 	metadata::ModelMetadata,
 	ortsys,
-	value::{Value, ValueType},
-	GraphOptimizationLevel
+	value::{Value, ValueType}
 };
-use crate::{environment::Environment, MemoryInfo, OperatorDomain};
 
+pub(crate) mod builder;
 pub(crate) mod input;
 pub(crate) mod output;
 pub use self::{
+	builder::{GraphOptimizationLevel, SessionBuilder},
 	input::{SessionInputValue, SessionInputs},
 	output::SessionOutputs
 };
-
-/// Creates a session using the builder pattern.
-///
-/// Once configured, use the [`SessionBuilder::commit_from_file`](crate::SessionBuilder::commit_from_file)
-/// method to 'commit' the builder configuration into a [`Session`].
-///
-/// ```
-/// # use ort::{GraphOptimizationLevel, Session};
-/// # fn main() -> ort::Result<()> {
-/// let session = Session::builder()?
-/// 	.with_optimization_level(GraphOptimizationLevel::Level1)?
-/// 	.with_intra_threads(1)?
-/// 	.commit_from_file("tests/data/upsample.onnx")?;
-/// # Ok(())
-/// # }
-/// ```
-pub struct SessionBuilder {
-	pub(crate) session_options_ptr: NonNull<ort_sys::OrtSessionOptions>,
-	memory_info: Option<Rc<MemoryInfo>>,
-	#[cfg(feature = "operator-libraries")]
-	custom_runtime_handles: Vec<*mut std::os::raw::c_void>,
-	operator_domains: Vec<Arc<OperatorDomain>>,
-	execution_providers: Vec<ExecutionProviderDispatch>
-}
-
-impl Clone for SessionBuilder {
-	fn clone(&self) -> Self {
-		let mut session_options_ptr = ptr::null_mut();
-		status_to_result(ortsys![unsafe CloneSessionOptions(self.session_options_ptr.as_ptr(), ptr::addr_of_mut!(session_options_ptr))])
-			.expect("error cloning session options");
-		assert_non_null_pointer(session_options_ptr, "OrtSessionOptions").unwrap();
-		Self {
-			session_options_ptr: unsafe { NonNull::new_unchecked(session_options_ptr) },
-			memory_info: self.memory_info.clone(),
-			#[cfg(feature = "operator-libraries")]
-			custom_runtime_handles: self.custom_runtime_handles.clone(),
-			operator_domains: self.operator_domains.clone(),
-			execution_providers: self.execution_providers.clone()
-		}
-	}
-}
-
-impl Drop for SessionBuilder {
-	fn drop(&mut self) {
-		#[cfg(feature = "operator-libraries")]
-		for &handle in &self.custom_runtime_handles {
-			close_lib_handle(handle);
-		}
-
-		ortsys![unsafe ReleaseSessionOptions(self.session_options_ptr.as_ptr())];
-	}
-}
-
-impl SessionBuilder {
-	/// Creates a new session builder.
-	///
-	/// ```
-	/// # use ort::{GraphOptimizationLevel, Session};
-	/// # fn main() -> ort::Result<()> {
-	/// let session = Session::builder()?
-	/// 	.with_optimization_level(GraphOptimizationLevel::Level1)?
-	/// 	.with_intra_threads(1)?
-	/// 	.commit_from_file("tests/data/upsample.onnx")?;
-	/// # Ok(())
-	/// # }
-	/// ```
-	pub fn new() -> Result<Self> {
-		let mut session_options_ptr: *mut ort_sys::OrtSessionOptions = std::ptr::null_mut();
-		ortsys![unsafe CreateSessionOptions(&mut session_options_ptr) -> Error::CreateSessionOptions; nonNull(session_options_ptr)];
-
-		Ok(Self {
-			session_options_ptr: unsafe { NonNull::new_unchecked(session_options_ptr) },
-			memory_info: None,
-			#[cfg(feature = "operator-libraries")]
-			custom_runtime_handles: Vec::new(),
-			operator_domains: Vec::new(),
-			execution_providers: Vec::new()
-		})
-	}
-
-	/// Registers a list of execution providers for this session. Execution providers are registered in the order they
-	/// are provided.
-	///
-	/// Execution providers will only work if the corresponding Cargo feature is enabled and ONNX Runtime was built
-	/// with support for the corresponding execution provider. Execution providers that do not have their corresponding
-	/// feature enabled will emit a warning.
-	///
-	/// ## Notes
-	///
-	/// - **Indiscriminate use of [`SessionBuilder::with_execution_providers`] in a library** (e.g. always enabling
-	///   `CUDAExecutionProvider`) **is discouraged** unless you allow the user to configure the execution providers by
-	///   providing a `Vec` of [`ExecutionProviderDispatch`]es.
-	pub fn with_execution_providers(self, execution_providers: impl IntoIterator<Item = ExecutionProviderDispatch>) -> Result<Self> {
-		apply_execution_providers(&self, execution_providers.into_iter());
-		Ok(self)
-	}
-
-	/// Configure the session to use a number of threads to parallelize the execution within nodes. If ONNX Runtime was
-	/// built with OpenMP (as is the case with Microsoft's prebuilt binaries), this will have no effect on the number of
-	/// threads used. Instead, you can configure the number of threads OpenMP uses via the `OMP_NUM_THREADS` environment
-	/// variable.
-	///
-	/// For configuring the number of threads used when the session execution mode is set to `Parallel`, see
-	/// [`SessionBuilder::with_inter_threads()`].
-	pub fn with_intra_threads(self, num_threads: usize) -> Result<Self> {
-		ortsys![unsafe SetIntraOpNumThreads(self.session_options_ptr.as_ptr(), num_threads as _) -> Error::CreateSessionOptions];
-		Ok(self)
-	}
-
-	/// Configure the session to use a number of threads to parallelize the execution of the graph. If nodes can be run
-	/// in parallel, this sets the maximum number of threads to use to run them in parallel.
-	///
-	/// This has no effect when the session execution mode is set to `Sequential`.
-	///
-	/// For configuring the number of threads used to parallelize the execution within nodes, see
-	/// [`SessionBuilder::with_intra_threads()`].
-	pub fn with_inter_threads(self, num_threads: usize) -> Result<Self> {
-		ortsys![unsafe SetInterOpNumThreads(self.session_options_ptr.as_ptr(), num_threads as _) -> Error::CreateSessionOptions];
-		Ok(self)
-	}
-
-	/// Enable/disable the parallel execution mode for this session. By default, this is disabled.
-	///
-	/// Parallel execution can improve performance for models with many branches, at the cost of higher memory usage.
-	/// You can configure the amount of threads used to parallelize the execution of the graph via
-	/// [`SessionBuilder::with_inter_threads()`].
-	pub fn with_parallel_execution(self, parallel_execution: bool) -> Result<Self> {
-		let execution_mode = if parallel_execution {
-			ort_sys::ExecutionMode::ORT_PARALLEL
-		} else {
-			ort_sys::ExecutionMode::ORT_SEQUENTIAL
-		};
-		ortsys![unsafe SetSessionExecutionMode(self.session_options_ptr.as_ptr(), execution_mode) -> Error::CreateSessionOptions];
-		Ok(self)
-	}
-
-	/// Set the session's optimization level. See [`GraphOptimizationLevel`] for more information on the different
-	/// optimization levels.
-	pub fn with_optimization_level(self, opt_level: GraphOptimizationLevel) -> Result<Self> {
-		ortsys![unsafe SetSessionGraphOptimizationLevel(self.session_options_ptr.as_ptr(), opt_level.into()) -> Error::CreateSessionOptions];
-		Ok(self)
-	}
-
-	/// After performing optimization (configurable with [`SessionBuilder::with_optimization_level`]), serializes the
-	/// newly optimized model to the given path (for 'offline' graph optimization).
-	///
-	/// Note that the file will only be created after the model is committed.
-	pub fn with_optimized_model_path<S: AsRef<str>>(self, path: S) -> Result<Self> {
-		#[cfg(windows)]
-		let path = path.as_ref().encode_utf16().chain([0]).collect::<Vec<_>>();
-		#[cfg(not(windows))]
-		let path = CString::new(path.as_ref())?;
-		ortsys![unsafe SetOptimizedModelFilePath(self.session_options_ptr.as_ptr(), path.as_ptr()) -> Error::CreateSessionOptions];
-		Ok(self)
-	}
-
-	/// Enables profiling. Profile information will be writen to `profiling_file` after profiling completes.
-	/// See [`Session::end_profiling`].
-	pub fn with_profiling<S: AsRef<str>>(self, profiling_file: S) -> Result<Self> {
-		#[cfg(windows)]
-		let profiling_file = profiling_file.as_ref().encode_utf16().chain([0]).collect::<Vec<_>>();
-		#[cfg(not(windows))]
-		let profiling_file = CString::new(profiling_file.as_ref())?;
-		ortsys![unsafe EnableProfiling(self.session_options_ptr.as_ptr(), profiling_file.as_ptr()) -> Error::CreateSessionOptions];
-		Ok(self)
-	}
-
-	/// Enables/disables memory pattern optimization. Disable it if the input size varies, i.e., dynamic batch
-	pub fn with_memory_pattern(self, enable: bool) -> Result<Self> {
-		if enable {
-			ortsys![unsafe EnableMemPattern(self.session_options_ptr.as_ptr()) -> Error::CreateSessionOptions];
-		} else {
-			ortsys![unsafe DisableMemPattern(self.session_options_ptr.as_ptr()) -> Error::CreateSessionOptions];
-		}
-		Ok(self)
-	}
-
-	/// Set the session's allocator options from a [`MemoryInfo`].
-	///
-	/// If not provided, the session is created using ONNX Runtime's default device allocator.
-	pub fn with_allocator(mut self, info: MemoryInfo) -> Result<Self> {
-		self.memory_info = Some(Rc::new(info));
-		Ok(self)
-	}
-
-	/// Registers a custom operator library at the given library path.
-	#[cfg(feature = "operator-libraries")]
-	#[cfg_attr(docsrs, doc(cfg(feature = "operator-libraries")))]
-	pub fn with_operator_library(mut self, lib_path: impl AsRef<str>) -> Result<Self> {
-		let path_cstr = CString::new(lib_path.as_ref())?;
-
-		let mut handle: *mut ::std::os::raw::c_void = std::ptr::null_mut();
-
-		let status = ortsys![unsafe RegisterCustomOpsLibrary(self.session_options_ptr.as_ptr(), path_cstr.as_ptr(), &mut handle)];
-
-		// per RegisterCustomOpsLibrary docs, release handle if there was an error and the handle
-		// is non-null
-		if let Err(e) = status_to_result(status).map_err(Error::CreateSessionOptions) {
-			if !handle.is_null() {
-				// handle was written to, should release it
-				close_lib_handle(handle);
-			}
-
-			return Err(e);
-		}
-
-		self.custom_runtime_handles.push(handle);
-
-		Ok(self)
-	}
-
-	/// Enables [`onnxruntime-extensions`](https://github.com/microsoft/onnxruntime-extensions) custom operators.
-	pub fn with_extensions(self) -> Result<Self> {
-		let status = ortsys![unsafe EnableOrtCustomOps(self.session_options_ptr.as_ptr())];
-		status_to_result(status).map_err(Error::CreateSessionOptions)?;
-		Ok(self)
-	}
-
-	pub fn with_operators(mut self, domain: impl Into<Arc<OperatorDomain>>) -> Result<Self> {
-		let domain = domain.into();
-		ortsys![unsafe AddCustomOpDomain(self.session_options_ptr.as_ptr(), domain.ptr()) -> Error::AddCustomOperatorDomain];
-		self.operator_domains.push(domain);
-		Ok(self)
-	}
-
-	/// Downloads a pre-trained ONNX model from the given URL and builds the session.
-	#[cfg(feature = "fetch-models")]
-	#[cfg_attr(docsrs, doc(cfg(feature = "fetch-models")))]
-	pub fn commit_from_url(self, model_url: impl AsRef<str>) -> Result<Session> {
-		let mut download_dir = ort_sys::internal::dirs::cache_dir()
-			.expect("could not determine cache directory")
-			.join("models");
-		if std::fs::create_dir_all(&download_dir).is_err() {
-			download_dir = std::env::current_dir().unwrap();
-		}
-
-		let url = model_url.as_ref();
-		let model_filename = PathBuf::from(url.split('/').last().unwrap());
-		let model_filepath = download_dir.join(model_filename);
-		let downloaded_path = if model_filepath.exists() {
-			tracing::info!(model_filepath = format!("{}", model_filepath.display()).as_str(), "Model already exists, skipping download");
-			model_filepath
-		} else {
-			tracing::info!(model_filepath = format!("{}", model_filepath.display()).as_str(), url = format!("{url:?}").as_str(), "Downloading model");
-
-			let resp = ureq::get(url).call().map_err(Box::new).map_err(FetchModelError::FetchError)?;
-
-			assert!(resp.has("Content-Length"));
-			let len = resp.header("Content-Length").and_then(|s| s.parse::<usize>().ok()).unwrap();
-			tracing::info!(len, "Downloading {} bytes", len);
-
-			let mut reader = resp.into_reader();
-
-			let f = std::fs::File::create(&model_filepath).unwrap();
-			let mut writer = std::io::BufWriter::new(f);
-
-			let bytes_io_count = std::io::copy(&mut reader, &mut writer).map_err(FetchModelError::IoError)?;
-			if bytes_io_count == len as u64 {
-				model_filepath
-			} else {
-				return Err(FetchModelError::CopyError {
-					expected: len as u64,
-					io: bytes_io_count
-				}
-				.into());
-			}
-		};
-
-		self.commit_from_file(downloaded_path)
-	}
-
-	/// Loads an ONNX model from a file and builds the session.
-	pub fn commit_from_file<P>(self, model_filepath_ref: P) -> Result<Session>
-	where
-		P: AsRef<Path>
-	{
-		let model_filepath = model_filepath_ref.as_ref();
-		if !model_filepath.exists() {
-			return Err(Error::FileDoesNotExist {
-				filename: model_filepath.to_path_buf()
-			});
-		}
-
-		// Build an OsString, then a vector of bytes to pass to C
-		let model_path = std::ffi::OsString::from(model_filepath);
-		#[cfg(target_family = "windows")]
-		let model_path: Vec<u16> = model_path
-            .encode_wide()
-            .chain(std::iter::once(0)) // Make sure we have a null terminated string
-            .collect();
-		#[cfg(not(target_family = "windows"))]
-		let model_path: Vec<std::os::raw::c_char> = model_path
-            .as_bytes()
-            .iter()
-            .chain(std::iter::once(&b'\0')) // Make sure we have a null terminated string
-            .map(|b| *b as std::os::raw::c_char)
-            .collect();
-
-		let env = get_environment()?;
-		apply_execution_providers(&self, env.execution_providers.iter().cloned());
-
-		if env.has_global_threadpool {
-			ortsys![unsafe DisablePerSessionThreads(self.session_options_ptr.as_ptr()) -> Error::CreateSessionOptions];
-		}
-
-		let env_ptr = env.env_ptr.load(Ordering::Relaxed);
-
-		let mut session_ptr: *mut ort_sys::OrtSession = std::ptr::null_mut();
-		ortsys![unsafe CreateSession(env_ptr, model_path.as_ptr(), self.session_options_ptr.as_ptr(), &mut session_ptr) -> Error::CreateSession; nonNull(session_ptr)];
-
-		let session_ptr = unsafe { NonNull::new_unchecked(session_ptr) };
-
-		let allocator = match &self.memory_info {
-			Some(info) => {
-				let mut allocator_ptr: *mut ort_sys::OrtAllocator = std::ptr::null_mut();
-				ortsys![unsafe CreateAllocator(session_ptr.as_ptr(), info.ptr.as_ptr(), &mut allocator_ptr) -> Error::CreateAllocator; nonNull(allocator_ptr)];
-				unsafe { Allocator::from_raw_unchecked(allocator_ptr) }
-			}
-			None => Allocator::default()
-		};
-
-		// Extract input and output properties
-		let num_input_nodes = dangerous::extract_inputs_count(session_ptr)?;
-		let num_output_nodes = dangerous::extract_outputs_count(session_ptr)?;
-		let inputs = (0..num_input_nodes)
-			.map(|i| dangerous::extract_input(session_ptr, &allocator, i))
-			.collect::<Result<Vec<Input>>>()?;
-		let outputs = (0..num_output_nodes)
-			.map(|i| dangerous::extract_output(session_ptr, &allocator, i))
-			.collect::<Result<Vec<Output>>>()?;
-
-		Ok(Session {
-			inner: Arc::new(SharedSessionInner {
-				session_ptr,
-				allocator,
-				_environment: Arc::clone(env)
-			}),
-			inputs,
-			outputs
-		})
-	}
-
-	/// Load an ONNX graph from memory and commit the session
-	/// For `.ort` models, we enable `session.use_ort_model_bytes_directly`.
-	/// For more information, check [Load ORT format model from an in-memory byte array](https://onnxruntime.ai/docs/performance/model-optimizations/ort-format-models.html#load-ort-format-model-from-an-in-memory-byte-array).
-	///
-	/// If you wish to store the model bytes and the [`InMemorySession`] in the same struct, look for crates that
-	/// facilitate creating self-referential structs, such as [`ouroboros`](https://github.com/joshua-maros/ouroboros).
-	pub fn commit_from_memory_directly(self, model_bytes: &[u8]) -> Result<InMemorySession<'_>> {
-		let str_to_char = |s: &str| {
-			s.as_bytes()
-				.iter()
-				.chain(std::iter::once(&b'\0')) // Make sure we have a null terminated string
-				.map(|b| *b as std::os::raw::c_char)
-				.collect::<Vec<std::os::raw::c_char>>()
-		};
-		// Enable zero-copy deserialization for models in `.ort` format.
-		ortsys![unsafe AddSessionConfigEntry(self.session_options_ptr.as_ptr(), str_to_char("session.use_ort_model_bytes_directly").as_ptr(), str_to_char("1").as_ptr())];
-		ortsys![unsafe AddSessionConfigEntry(self.session_options_ptr.as_ptr(), str_to_char("session.use_ort_model_bytes_for_initializers").as_ptr(), str_to_char("1").as_ptr())];
-
-		let session = self.commit_from_memory(model_bytes)?;
-
-		Ok(InMemorySession { session, phantom: PhantomData })
-	}
-
-	/// Load an ONNX graph from memory and commit the session.
-	pub fn commit_from_memory(self, model_bytes: &[u8]) -> Result<Session> {
-		let mut session_ptr: *mut ort_sys::OrtSession = std::ptr::null_mut();
-
-		let env = get_environment()?;
-		apply_execution_providers(&self, env.execution_providers.iter().cloned());
-
-		if env.has_global_threadpool {
-			ortsys![unsafe DisablePerSessionThreads(self.session_options_ptr.as_ptr()) -> Error::CreateSessionOptions];
-		}
-
-		let env_ptr = env.env_ptr.load(Ordering::Relaxed);
-
-		let model_data = model_bytes.as_ptr().cast::<std::ffi::c_void>();
-		let model_data_length = model_bytes.len();
-		ortsys![
-			unsafe CreateSessionFromArray(env_ptr, model_data, model_data_length as _, self.session_options_ptr.as_ptr(), &mut session_ptr) -> Error::CreateSession;
-			nonNull(session_ptr)
-		];
-
-		let session_ptr = unsafe { NonNull::new_unchecked(session_ptr) };
-
-		let allocator = match &self.memory_info {
-			Some(info) => {
-				let mut allocator_ptr: *mut ort_sys::OrtAllocator = std::ptr::null_mut();
-				ortsys![unsafe CreateAllocator(session_ptr.as_ptr(), info.ptr.as_ptr(), &mut allocator_ptr) -> Error::CreateAllocator; nonNull(allocator_ptr)];
-				unsafe { Allocator::from_raw_unchecked(allocator_ptr) }
-			}
-			None => Allocator::default()
-		};
-
-		// Extract input and output properties
-		let num_input_nodes = dangerous::extract_inputs_count(session_ptr)?;
-		let num_output_nodes = dangerous::extract_outputs_count(session_ptr)?;
-		let inputs = (0..num_input_nodes)
-			.map(|i| dangerous::extract_input(session_ptr, &allocator, i))
-			.collect::<Result<Vec<Input>>>()?;
-		let outputs = (0..num_output_nodes)
-			.map(|i| dangerous::extract_output(session_ptr, &allocator, i))
-			.collect::<Result<Vec<Output>>>()?;
-
-		let session = Session {
-			inner: Arc::new(SharedSessionInner {
-				session_ptr,
-				allocator,
-				_environment: Arc::clone(env)
-			}),
-			inputs,
-			outputs
-		};
-		Ok(session)
-	}
-}
 
 /// Holds onto an [`ort_sys::OrtSession`] pointer and its associated allocator.
 ///
@@ -829,19 +393,6 @@ unsafe impl Send for Session {}
 // temporary bug in ONNX Runtime or a wontfix. Maybe this impl should be removed just to be safe?
 unsafe impl Sync for Session {}
 
-#[cfg(all(unix, feature = "operator-libraries"))]
-fn close_lib_handle(handle: *mut std::os::raw::c_void) {
-	unsafe { libc::dlclose(handle) };
-}
-
-#[cfg(all(windows, feature = "operator-libraries"))]
-fn close_lib_handle(handle: *mut std::os::raw::c_void) {
-	unsafe { winapi::um::libloaderapi::FreeLibrary(handle as winapi::shared::minwindef::HINSTANCE) };
-}
-
-/// This module contains dangerous functions working on raw pointers.
-/// Those functions are only to be used from inside the
-/// `SessionBuilder::commit_from_file()` method.
 mod dangerous {
 	use super::*;
 	use crate::value::{extract_data_type_from_map_info, extract_data_type_from_sequence_info, extract_data_type_from_tensor_info};
